+++ Backend/src/social_media_platforms/facebook/facebook.interface.ts
@@ -1,0 +1,19 @@
+export interface FacebookPage {
+  id: string;
+  name: string;
+  access_token: string;
+  picture?: {
+    data?: {
+      url?: string;
+    };
+  };
+}
+
+export interface FacebookAttachedMedia {
+  media_fbid: string;
+}
+
+export interface FacebookPostResponse {
+  id: string;
+  post_id?: string;
+}
