+++ Backend/src/social_media_platforms/facebook/facebook.module.ts
@@ -3,13 +3,15 @@ import { Module } from '@nestjs/common';
 import { FacebookController } from './facebook.controller';
 import { FacebookService } from './facebook.service';
 import { PrismaModule } from 'src/prisma/prisma.module';
+import { FacebookAuthService } from './facebook-auth.service';
+import { FacebookMediaService } from './facebook-media.service';
 
 @Module({
   imports: [
     PrismaModule,
   ],
   controllers: [FacebookController],
-  providers: [FacebookService],
+  providers: [FacebookService, FacebookAuthService, FacebookMediaService],
   exports: [FacebookService],
 })
 export class FacebookModule {}
